+++ src/main/scala/sc4pac/api/message.scala
@@ -158,6 +158,7 @@ object PromptMessage {
   case class DownloadFailedSelectMirror(
     url: java.net.URI,
     reason: ErrorMessage.DownloadFailed,
+    promptForSimtropolisToken: Boolean,
     choices: Seq[String],
     token: String,
     responses: Map[String, ResponseMessage],
@@ -165,15 +166,16 @@ object PromptMessage {
     override def accept(response: ResponseMessage): Boolean = response.token == token
   }
   object DownloadFailedSelectMirror {
-    def apply(url: java.net.URI, reason: ErrorMessage.DownloadFailed): DownloadFailedSelectMirror = {
+    def apply(url: java.net.URI, reason: ErrorMessage.DownloadFailed, promptForSimtropolisToken: Boolean): DownloadFailedSelectMirror = {
       val token = scala.util.Random.nextInt().toHexString
       DownloadFailedSelectMirror(
         url = url,
         reason = reason,
+        promptForSimtropolisToken = promptForSimtropolisToken,
         choices = Seq("Retry", "Cancel"),
         token = token,
         responses = Map(
-          "Retry" -> ResponseMessage(token, ujson.Obj("retry" -> true, "localMirror" -> ujson.Arr())),
+          "Retry" -> ResponseMessage(token, ujson.Obj("retry" -> true, "localMirror" -> ujson.Arr(), "simtropolisToken" -> ujson.Arr())),
           // Select mirror -> retry=true, localMirror=[path]  (must be constructed manually)
           "Cancel" -> ResponseMessage(token, ujson.Obj("retry" -> false)),
         ),
@@ -183,6 +185,7 @@ object PromptMessage {
     case class ResponseData(
       retry: Boolean,
       localMirror: Option[java.nio.file.Path] = None,
+      simtropolisToken: Option[String] = None,
     ) derives UP.ReadWriter
   }