+++ src/main/scala/sc4pac/error.scala
@@ -30,7 +30,7 @@ final class ExtractionFailed(val title: String, val detail: String) extends java
   def withDetail(newDetail: String): ExtractionFailed = ExtractionFailed(title, newDetail)
 }
 
-final class DownloadFailed(val title: String, val detail: String, val url: Option[java.net.URI]) extends java.io.IOException(s"$title $detail") with Sc4pacErr
+final class DownloadFailed(val title: String, val detail: String, val url: Option[java.net.URI], val promptForSimtropolisToken: Boolean = false) extends java.io.IOException(s"$title $detail") with Sc4pacErr
 
 final class ChecksumError(val title: String, val detail: String) extends java.io.IOException(s"$title $detail") with Sc4pacErr