+++ CHANGELOG.md
@@ -9,14 +9,15 @@
 - an issue preventing the import and export of JSON mod sets if the JSON file was too large ([#39][gui39])
 - Extraction failures now display the name of the originating package (#42)
 
-[gui39]: https://github.com/memo33/sc4pac-gui/issues/39
-
 ### Changed
 - API upgraded to 2.8:
   - added optional `includePlugins` parameter to `/profiles.list` endpoint
   - added `/plugins.reinstall` endpoint and corresponding installation status
   - added `/profiles.remove` and `/profiles.rename` endpoints
+  - changed format of message `/prompt/json/update/download-failed-select-mirror` to allow setting a Simtropolis token ([#35][gui35])
 
+[gui35]: https://github.com/memo33/sc4pac-gui/issues/35
+[gui39]: https://github.com/memo33/sc4pac-gui/issues/39
 
 ## [0.8.1] - 2025-09-20
 ### Added